@@ -1,1032 +1,803 @@
 <?php
-/**
- * CodeIgniter
- *
- * An open source application development framework for PHP
- *
- * This content is released under the MIT License (MIT)
- *
- * Copyright (c) 2014 - 2016, British Columbia Institute of Technology
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- * @package	CodeIgniter
- * @author	EllisLab Dev Team
- * @copyright	Copyright (c) 2008 - 2014, EllisLab, Inc. (https://ellislab.com/)
- * @copyright	Copyright (c) 2014 - 2016, British Columbia Institute of Technology (http://bcit.ca/)
- * @license	http://opensource.org/licenses/MIT	MIT License
- * @link	https://codeigniter.com
- * @since	Version 1.0.0
- * @filesource
- */
-defined('BASEPATH') OR exit('No direct script access allowed');
 
 /**
- * CodeIgniter Form Helpers
+ * This file is part of CodeIgniter 4 framework.
  *
- * @package		CodeIgniter
- * @subpackage	Helpers
- * @category	Helpers
- * @author		EllisLab Dev Team
- * @link		https://codeigniter.com/user_guide/helpers/form_helper.html
+ * (c) CodeIgniter Foundation <admin@codeigniter.com>
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
  */
 
-// ------------------------------------------------------------------------
+use CodeIgniter\Validation\Exceptions\ValidationException;
+use Config\App;
+use Config\Services;
+use Config\Validation;
 
-if ( ! function_exists('form_open'))
-{
-	/**
-	 * Form Declaration
-	 *
-	 * Creates the opening portion of the form.
-	 *
-	 * @param	string	the URI segments of the form destination
-	 * @param	array	a key/value pair of attributes
-	 * @param	array	a key/value pair hidden data
-	 * @return	string
-	 */
-	function form_open($action = '', $attributes = array(), $hidden = array())
-	{
-		$CI =& get_instance();
+// CodeIgniter Form Helpers
 
-		// If no action is provided then set to the current url
-		if ( ! $action)
-		{
-			$action = $CI->config->site_url($CI->uri->uri_string());
-		}
-		// If an action is not a full URL then turn it into one
-		elseif (strpos($action, '://') === FALSE)
-		{
-			$action = $CI->config->site_url($action);
-		}
+if (! function_exists('form_open')) {
+    /**
+     * Form Declaration
+     *
+     * Creates the opening portion of the form.
+     *
+     * @param string       $action     the URI segments of the form destination
+     * @param array|string $attributes a key/value pair of attributes, or string representation
+     * @param array        $hidden     a key/value pair hidden data
+     */
+    function form_open(string $action = '', $attributes = [], array $hidden = []): string
+    {
+        // If no action is provided then set to the current url
+        if (! $action) {
+            $action = current_url(true);
+        } // If an action is not a full URL then turn it into one
+        elseif (strpos($action, '://') === false) {
+            // If an action has {locale}
+            if (strpos($action, '{locale}') !== false) {
+                $action = str_replace('{locale}', Services::request()->getLocale(), $action);
+            }
 
-		$attributes = _attributes_to_string($attributes);
+            $action = site_url($action);
+        }
 
-		if (stripos($attributes, 'method=') === FALSE)
-		{
-			$attributes .= ' method="post"';
-		}
+        if (is_array($attributes) && array_key_exists('csrf_id', $attributes)) {
+            $csrfId = $attributes['csrf_id'];
+            unset($attributes['csrf_id']);
+        }
 
-		if (stripos($attributes, 'accept-charset=') === FALSE)
-		{
-			$attributes .= ' accept-charset="'.strtolower(config_item('charset')).'"';
-		}
+        $attributes = stringify_attributes($attributes);
 
-		$form = '<form action="'.$action.'"'.$attributes.">\n";
+        if (stripos($attributes, 'method=') === false) {
+            $attributes .= ' method="post"';
+        }
+        if (stripos($attributes, 'accept-charset=') === false) {
+            $config = config(App::class);
+            $attributes .= ' accept-charset="' . strtolower($config->charset) . '"';
+        }
 
-		// Add CSRF field if enabled, but leave it out for GET requests and requests to external websites
-		if ($CI->config->item('csrf_protection') === TRUE && strpos($action, $CI->config->base_url()) !== FALSE && ! stripos($form, 'method="get"'))
-		{
-			$hidden[$CI->security->get_csrf_token_name()] = $CI->security->get_csrf_hash();
-		}
+        $form = '<form action="' . $action . '"' . $attributes . ">\n";
 
-		if (is_array($hidden))
-		{
-			foreach ($hidden as $name => $value)
-			{
-				$form .= '<input type="hidden" name="'.$name.'" value="'.html_escape($value).'" style="display:none;" />'."\n";
-			}
-		}
+        // Add CSRF field if enabled, but leave it out for GET requests and requests to external websites
+        $before = Services::filters()->getFilters()['before'];
 
-		return $form;
-	}
+        if ((in_array('csrf', $before, true) || array_key_exists('csrf', $before)) && strpos($action, base_url()) !== false && ! stripos($form, 'method="get"')) {
+            $form .= csrf_field($csrfId ?? null);
+        }
+
+        foreach ($hidden as $name => $value) {
+            $form .= form_hidden($name, $value);
+        }
+
+        return $form;
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_open_multipart')) {
+    /**
+     * Form Declaration - Multipart type
+     *
+     * Creates the opening portion of the form, but with "multipart/form-data".
+     *
+     * @param string       $action     The URI segments of the form destination
+     * @param array|string $attributes A key/value pair of attributes, or the same as a string
+     * @param array        $hidden     A key/value pair hidden data
+     */
+    function form_open_multipart(string $action = '', $attributes = [], array $hidden = []): string
+    {
+        if (is_string($attributes)) {
+            $attributes .= ' enctype="' . esc('multipart/form-data') . '"';
+        } else {
+            $attributes['enctype'] = 'multipart/form-data';
+        }
 
-if ( ! function_exists('form_open_multipart'))
-{
-	/**
-	 * Form Declaration - Multipart type
-	 *
-	 * Creates the opening portion of the form, but with "multipart/form-data".
-	 *
-	 * @param	string	the URI segments of the form destination
-	 * @param	array	a key/value pair of attributes
-	 * @param	array	a key/value pair hidden data
-	 * @return	string
-	 */
-	function form_open_multipart($action = '', $attributes = array(), $hidden = array())
-	{
-		if (is_string($attributes))
-		{
-			$attributes .= ' enctype="multipart/form-data"';
-		}
-		else
-		{
-			$attributes['enctype'] = 'multipart/form-data';
-		}
-
-		return form_open($action, $attributes, $hidden);
-	}
+        return form_open($action, $attributes, $hidden);
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_hidden')) {
+    /**
+     * Hidden Input Field
+     *
+     * Generates hidden fields. You can pass a simple key/value string or
+     * an associative array with multiple values.
+     *
+     * @param array|string $name  Field name or associative array to create multiple fields
+     * @param array|string $value Field value
+     */
+    function form_hidden($name, $value = '', bool $recursing = false): string
+    {
+        static $form;
 
-if ( ! function_exists('form_hidden'))
-{
-	/**
-	 * Hidden Input Field
-	 *
-	 * Generates hidden fields. You can pass a simple key/value string or
-	 * an associative array with multiple values.
-	 *
-	 * @param	mixed	$name		Field name
-	 * @param	string	$value		Field value
-	 * @param	bool	$recursing
-	 * @return	string
-	 */
-	function form_hidden($name, $value = '', $recursing = FALSE)
-	{
-		static $form;
+        if ($recursing === false) {
+            $form = "\n";
+        }
 
-		if ($recursing === FALSE)
-		{
-			$form = "\n";
-		}
+        if (is_array($name)) {
+            foreach ($name as $key => $val) {
+                form_hidden($key, $val, true);
+            }
 
-		if (is_array($name))
-		{
-			foreach ($name as $key => $val)
-			{
-				form_hidden($key, $val, TRUE);
-			}
+            return $form;
+        }
 
-			return $form;
-		}
+        if (! is_array($value)) {
+            $form .= form_input($name, $value, '', 'hidden');
+        } else {
+            foreach ($value as $k => $v) {
+                $k = is_int($k) ? '' : $k;
+                form_hidden($name . '[' . $k . ']', $v, true);
+            }
+        }
 
-		if ( ! is_array($value))
-		{
-			$form .= '<input type="hidden" name="'.$name.'" value="'.html_escape($value)."\" />\n";
-		}
-		else
-		{
-			foreach ($value as $k => $v)
-			{
-				$k = is_int($k) ? '' : $k;
-				form_hidden($name.'['.$k.']', $v, TRUE);
-			}
-		}
-
-		return $form;
-	}
+        return $form;
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_input')) {
+    /**
+     * Text Input Field. If 'type' is passed in the $type field, it will be
+     * used as the input type, for making 'email', 'phone', etc input fields.
+     *
+     * @param array|string        $data
+     * @param array|object|string $extra string, array, object that can be cast to array
+     */
+    function form_input($data = '', string $value = '', $extra = '', string $type = 'text'): string
+    {
+        $defaults = [
+            'type'  => $type,
+            'name'  => is_array($data) ? '' : $data,
+            'value' => $value,
+        ];
 
-if ( ! function_exists('form_input'))
-{
-	/**
-	 * Text Input Field
-	 *
-	 * @param	mixed
-	 * @param	string
-	 * @param	mixed
-	 * @return	string
-	 */
-	function form_input($data = '', $value = '', $extra = '')
-	{
-		$defaults = array(
-			'type' => 'text',
-			'name' => is_array($data) ? '' : $data,
-			'value' => $value
-		);
-
-		return '<input '._parse_form_attributes($data, $defaults)._attributes_to_string($extra)." />\n";
-	}
+        return '<input ' . parse_form_attributes($data, $defaults) . stringify_attributes($extra) . _solidus() . ">\n";
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_password')) {
+    /**
+     * Password Field
+     *
+     * Identical to the input function but adds the "password" type
+     *
+     * @param array|string        $data
+     * @param array|object|string $extra string, array, object that can be cast to array
+     */
+    function form_password($data = '', string $value = '', $extra = ''): string
+    {
+        if (! is_array($data)) {
+            $data = ['name' => $data];
+        }
+        $data['type'] = 'password';
 
-if ( ! function_exists('form_password'))
-{
-	/**
-	 * Password Field
-	 *
-	 * Identical to the input function but adds the "password" type
-	 *
-	 * @param	mixed
-	 * @param	string
-	 * @param	mixed
-	 * @return	string
-	 */
-	function form_password($data = '', $value = '', $extra = '')
-	{
-		is_array($data) OR $data = array('name' => $data);
-		$data['type'] = 'password';
-		return form_input($data, $value, $extra);
-	}
+        return form_input($data, $value, $extra);
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_upload')) {
+    /**
+     * Upload Field
+     *
+     * Identical to the input function but adds the "file" type
+     *
+     * @param array|string        $data
+     * @param array|object|string $extra string, array, object that can be cast to array
+     */
+    function form_upload($data = '', string $value = '', $extra = ''): string
+    {
+        $defaults = [
+            'type' => 'file',
+            'name' => '',
+        ];
 
-if ( ! function_exists('form_upload'))
-{
-	/**
-	 * Upload Field
-	 *
-	 * Identical to the input function but adds the "file" type
-	 *
-	 * @param	mixed
-	 * @param	string
-	 * @param	mixed
-	 * @return	string
-	 */
-	function form_upload($data = '', $value = '', $extra = '')
-	{
-		$defaults = array('type' => 'file', 'name' => '');
-		is_array($data) OR $data = array('name' => $data);
-		$data['type'] = 'file';
+        if (! is_array($data)) {
+            $data = ['name' => $data];
+        }
 
-		return '<input '._parse_form_attributes($data, $defaults)._attributes_to_string($extra)." />\n";
-	}
+        $data['type'] = 'file';
+
+        return '<input ' . parse_form_attributes($data, $defaults) . stringify_attributes($extra) . _solidus() . ">\n";
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_textarea')) {
+    /**
+     * Textarea field
+     *
+     * @param array|string        $data
+     * @param array|object|string $extra string, array, object that can be cast to array
+     */
+    function form_textarea($data = '', string $value = '', $extra = ''): string
+    {
+        $defaults = [
+            'name' => is_array($data) ? '' : $data,
+            'cols' => '40',
+            'rows' => '10',
+        ];
+        if (! is_array($data) || ! isset($data['value'])) {
+            $val = $value;
+        } else {
+            $val = $data['value'];
+            unset($data['value']); // textareas don't use the value attribute
+        }
 
-if ( ! function_exists('form_textarea'))
-{
-	/**
-	 * Textarea field
-	 *
-	 * @param	mixed	$data
-	 * @param	string	$value
-	 * @param	mixed	$extra
-	 * @return	string
-	 */
-	function form_textarea($data = '', $value = '', $extra = '')
-	{
-		$defaults = array(
-			'name' => is_array($data) ? '' : $data,
-			'cols' => '40',
-			'rows' => '10'
-		);
+        // Unsets default rows and cols if defined in extra field as array or string.
+        if ((is_array($extra) && array_key_exists('rows', $extra)) || (is_string($extra) && stripos(preg_replace('/\s+/', '', $extra), 'rows=') !== false)) {
+            unset($defaults['rows']);
+        }
 
-		if ( ! is_array($data) OR ! isset($data['value']))
-		{
-			$val = $value;
-		}
-		else
-		{
-			$val = $data['value'];
-			unset($data['value']); // textareas don't use the value attribute
-		}
+        if ((is_array($extra) && array_key_exists('cols', $extra)) || (is_string($extra) && stripos(preg_replace('/\s+/', '', $extra), 'cols=') !== false)) {
+            unset($defaults['cols']);
+        }
 
-		return '<textarea '._parse_form_attributes($data, $defaults)._attributes_to_string($extra).'>'
-			.html_escape($val)
-			."</textarea>\n";
-	}
+        return '<textarea ' . rtrim(parse_form_attributes($data, $defaults)) . stringify_attributes($extra) . '>'
+                . htmlspecialchars($val)
+                . "</textarea>\n";
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_multiselect')) {
+    /**
+     * Multi-select menu
+     *
+     * @param array|string        $name
+     * @param array|object|string $extra string, array, object that can be cast to array
+     */
+    function form_multiselect($name = '', array $options = [], array $selected = [], $extra = ''): string
+    {
+        $extra = stringify_attributes($extra);
 
-if ( ! function_exists('form_multiselect'))
-{
-	/**
-	 * Multi-select menu
-	 *
-	 * @param	string
-	 * @param	array
-	 * @param	mixed
-	 * @param	mixed
-	 * @return	string
-	 */
-	function form_multiselect($name = '', $options = array(), $selected = array(), $extra = '')
-	{
-		$extra = _attributes_to_string($extra);
-		if (stripos($extra, 'multiple') === FALSE)
-		{
-			$extra .= ' multiple="multiple"';
-		}
+        if (stripos($extra, 'multiple') === false) {
+            $extra .= ' multiple="multiple"';
+        }
 
-		return form_dropdown($name, $options, $selected, $extra);
-	}
+        return form_dropdown($name, $options, $selected, $extra);
+    }
 }
 
-// --------------------------------------------------------------------
+if (! function_exists('form_dropdown')) {
+    /**
+     * Drop-down Menu
+     *
+     * @param array|string        $data
+     * @param array|string        $options
+     * @param array|string        $selected
+     * @param array|object|string $extra    string, array, object that can be cast to array
+     */
+    function form_dropdown($data = '', $options = [], $selected = [], $extra = ''): string
+    {
+        $defaults = [];
+        if (is_array($data)) {
+            if (isset($data['selected'])) {
+                $selected = $data['selected'];
+                unset($data['selected']); // select tags don't have a selected attribute
+            }
+            if (isset($data['options'])) {
+                $options = $data['options'];
+                unset($data['options']); // select tags don't use an options attribute
+            }
+        } else {
+            $defaults = ['name' => $data];
+        }
 
-if ( ! function_exists('form_dropdown'))
-{
-	/**
-	 * Drop-down Menu
-	 *
-	 * @param	mixed	$data
-	 * @param	mixed	$options
-	 * @param	mixed	$selected
-	 * @param	mixed	$extra
-	 * @return	string
-	 */
-	function form_dropdown($data = '', $options = array(), $selected = array(), $extra = '')
-	{
-		$defaults = array();
+        if (! is_array($selected)) {
+            $selected = [$selected];
+        }
+        if (! is_array($options)) {
+            $options = [$options];
+        }
 
-		if (is_array($data))
-		{
-			if (isset($data['selected']))
-			{
-				$selected = $data['selected'];
-				unset($data['selected']); // select tags don't have a selected attribute
-			}
+        // If no selected state was submitted we will attempt to set it automatically
+        if ($selected === []) {
+            if (is_array($data)) {
+                if (isset($data['name'], $_POST[$data['name']])) {
+                    $selected = [$_POST[$data['name']]];
+                }
+            } elseif (isset($_POST[$data])) {
+                $selected = [$_POST[$data]];
+            }
+        }
 
-			if (isset($data['options']))
-			{
-				$options = $data['options'];
-				unset($data['options']); // select tags don't use an options attribute
-			}
-		}
-		else
-		{
-			$defaults = array('name' => $data);
-		}
+        // Standardize selected as strings, like the option keys will be
+        foreach ($selected as $key => $item) {
+            $selected[$key] = (string) $item;
+        }
 
-		is_array($selected) OR $selected = array($selected);
-		is_array($options) OR $options = array($options);
+        $extra    = stringify_attributes($extra);
+        $multiple = (count($selected) > 1 && stripos($extra, 'multiple') === false) ? ' multiple="multiple"' : '';
+        $form     = '<select ' . rtrim(parse_form_attributes($data, $defaults)) . $extra . $multiple . ">\n";
 
-		// If no selected state was submitted we will attempt to set it automatically
-		if (empty($selected))
-		{
-			if (is_array($data))
-			{
-				if (isset($data['name'], $_POST[$data['name']]))
-				{
-					$selected = array($_POST[$data['name']]);
-				}
-			}
-			elseif (isset($_POST[$data]))
-			{
-				$selected = array($_POST[$data]);
-			}
-		}
+        foreach ($options as $key => $val) {
+            // Keys should always be strings for strict comparison
+            $key = (string) $key;
 
-		$extra = _attributes_to_string($extra);
+            if (is_array($val)) {
+                if ($val === []) {
+                    continue;
+                }
 
-		$multiple = (count($selected) > 1 && stripos($extra, 'multiple') === FALSE) ? ' multiple="multiple"' : '';
+                $form .= '<optgroup label="' . $key . "\">\n";
 
-		$form = '<select '.rtrim(_parse_form_attributes($data, $defaults)).$extra.$multiple.">\n";
+                foreach ($val as $optgroupKey => $optgroupVal) {
+                    // Keys should always be strings for strict comparison
+                    $optgroupKey = (string) $optgroupKey;
 
-		foreach ($options as $key => $val)
-		{
-			$key = (string) $key;
+                    $sel = in_array($optgroupKey, $selected, true) ? ' selected="selected"' : '';
+                    $form .= '<option value="' . htmlspecialchars($optgroupKey) . '"' . $sel . '>' . $optgroupVal . "</option>\n";
+                }
 
-			if (is_array($val))
-			{
-				if (empty($val))
-				{
-					continue;
-				}
+                $form .= "</optgroup>\n";
+            } else {
+                $form .= '<option value="' . htmlspecialchars($key) . '"'
+                    . (in_array($key, $selected, true) ? ' selected="selected"' : '') . '>'
+                    . $val . "</option>\n";
+            }
+        }
 
-				$form .= '<optgroup label="'.$key."\">\n";
-
-				foreach ($val as $optgroup_key => $optgroup_val)
-				{
-					$sel = in_array($optgroup_key, $selected) ? ' selected="selected"' : '';
-					$form .= '<option value="'.html_escape($optgroup_key).'"'.$sel.'>'
-						.(string) $optgroup_val."</option>\n";
-				}
-
-				$form .= "</optgroup>\n";
-			}
-			else
-			{
-				$form .= '<option value="'.html_escape($key).'"'
-					.(in_array($key, $selected) ? ' selected="selected"' : '').'>'
-					.(string) $val."</option>\n";
-			}
-		}
-
-		return $form."</select>\n";
-	}
+        return $form . "</select>\n";
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_checkbox')) {
+    /**
+     * Checkbox Field
+     *
+     * @param array|string        $data
+     * @param array|object|string $extra string, array, object that can be cast to array
+     */
+    function form_checkbox($data = '', string $value = '', bool $checked = false, $extra = ''): string
+    {
+        $defaults = [
+            'type'  => 'checkbox',
+            'name'  => (! is_array($data) ? $data : ''),
+            'value' => $value,
+        ];
 
-if ( ! function_exists('form_checkbox'))
-{
-	/**
-	 * Checkbox Field
-	 *
-	 * @param	mixed
-	 * @param	string
-	 * @param	bool
-	 * @param	mixed
-	 * @return	string
-	 */
-	function form_checkbox($data = '', $value = '', $checked = FALSE, $extra = '')
-	{
-		$defaults = array('type' => 'checkbox', 'name' => ( ! is_array($data) ? $data : ''), 'value' => $value);
+        if (is_array($data) && array_key_exists('checked', $data)) {
+            $checked = $data['checked'];
 
-		if (is_array($data) && array_key_exists('checked', $data))
-		{
-			$checked = $data['checked'];
+            if ($checked === false) {
+                unset($data['checked']);
+            } else {
+                $data['checked'] = 'checked';
+            }
+        }
 
-			if ($checked == FALSE)
-			{
-				unset($data['checked']);
-			}
-			else
-			{
-				$data['checked'] = 'checked';
-			}
-		}
+        if ($checked === true) {
+            $defaults['checked'] = 'checked';
+        }
 
-		if ($checked == TRUE)
-		{
-			$defaults['checked'] = 'checked';
-		}
-		else
-		{
-			unset($defaults['checked']);
-		}
-
-		return '<input '._parse_form_attributes($data, $defaults)._attributes_to_string($extra)." />\n";
-	}
+        return '<input ' . parse_form_attributes($data, $defaults) . stringify_attributes($extra) . _solidus() . ">\n";
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_radio')) {
+    /**
+     * Radio Button
+     *
+     * @param array|string        $data
+     * @param array|object|string $extra string, array, object that can be cast to array
+     */
+    function form_radio($data = '', string $value = '', bool $checked = false, $extra = ''): string
+    {
+        if (! is_array($data)) {
+            $data = ['name' => $data];
+        }
+        $data['type'] = 'radio';
 
-if ( ! function_exists('form_radio'))
-{
-	/**
-	 * Radio Button
-	 *
-	 * @param	mixed
-	 * @param	string
-	 * @param	bool
-	 * @param	mixed
-	 * @return	string
-	 */
-	function form_radio($data = '', $value = '', $checked = FALSE, $extra = '')
-	{
-		is_array($data) OR $data = array('name' => $data);
-		$data['type'] = 'radio';
-
-		return form_checkbox($data, $value, $checked, $extra);
-	}
+        return form_checkbox($data, $value, $checked, $extra);
+    }
 }
 
-// ------------------------------------------------------------------------
-
-if ( ! function_exists('form_submit'))
-{
-	/**
-	 * Submit Button
-	 *
-	 * @param	mixed
-	 * @param	string
-	 * @param	mixed
-	 * @return	string
-	 */
-	function form_submit($data = '', $value = '', $extra = '')
-	{
-		$defaults = array(
-			'type' => 'submit',
-			'name' => is_array($data) ? '' : $data,
-			'value' => $value
-		);
-
-		return '<input '._parse_form_attributes($data, $defaults)._attributes_to_string($extra)." />\n";
-	}
+if (! function_exists('form_submit')) {
+    /**
+     * Submit Button
+     *
+     * @param array|string        $data
+     * @param array|object|string $extra string, array, object that can be cast to array
+     */
+    function form_submit($data = '', string $value = '', $extra = ''): string
+    {
+        return form_input($data, $value, $extra, 'submit');
+    }
 }
 
-// ------------------------------------------------------------------------
-
-if ( ! function_exists('form_reset'))
-{
-	/**
-	 * Reset Button
-	 *
-	 * @param	mixed
-	 * @param	string
-	 * @param	mixed
-	 * @return	string
-	 */
-	function form_reset($data = '', $value = '', $extra = '')
-	{
-		$defaults = array(
-			'type' => 'reset',
-			'name' => is_array($data) ? '' : $data,
-			'value' => $value
-		);
-
-		return '<input '._parse_form_attributes($data, $defaults)._attributes_to_string($extra)." />\n";
-	}
+if (! function_exists('form_reset')) {
+    /**
+     * Reset Button
+     *
+     * @param array|string        $data
+     * @param array|object|string $extra string, array, object that can be cast to array
+     */
+    function form_reset($data = '', string $value = '', $extra = ''): string
+    {
+        return form_input($data, $value, $extra, 'reset');
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_button')) {
+    /**
+     * Form Button
+     *
+     * @param array|string        $data
+     * @param array|object|string $extra string, array, object that can be cast to array
+     */
+    function form_button($data = '', string $content = '', $extra = ''): string
+    {
+        $defaults = [
+            'name' => is_array($data) ? '' : $data,
+            'type' => 'button',
+        ];
 
-if ( ! function_exists('form_button'))
-{
-	/**
-	 * Form Button
-	 *
-	 * @param	mixed
-	 * @param	string
-	 * @param	mixed
-	 * @return	string
-	 */
-	function form_button($data = '', $content = '', $extra = '')
-	{
-		$defaults = array(
-			'name' => is_array($data) ? '' : $data,
-			'type' => 'button'
-		);
+        if (is_array($data) && isset($data['content'])) {
+            $content = $data['content'];
+            unset($data['content']); // content is not an attribute
+        }
 
-		if (is_array($data) && isset($data['content']))
-		{
-			$content = $data['content'];
-			unset($data['content']); // content is not an attribute
-		}
-
-		return '<button '._parse_form_attributes($data, $defaults)._attributes_to_string($extra).'>'
-			.$content
-			."</button>\n";
-	}
+        return '<button ' . parse_form_attributes($data, $defaults) . stringify_attributes($extra) . '>'
+                . $content
+                . "</button>\n";
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_label')) {
+    /**
+     * Form Label Tag
+     *
+     * @param string $labelText  The text to appear onscreen
+     * @param string $id         The id the label applies to
+     * @param array  $attributes Additional attributes
+     */
+    function form_label(string $labelText = '', string $id = '', array $attributes = []): string
+    {
+        $label = '<label';
 
-if ( ! function_exists('form_label'))
-{
-	/**
-	 * Form Label Tag
-	 *
-	 * @param	string	The text to appear onscreen
-	 * @param	string	The id the label applies to
-	 * @param	string	Additional attributes
-	 * @return	string
-	 */
-	function form_label($label_text = '', $id = '', $attributes = array())
-	{
+        if ($id !== '') {
+            $label .= ' for="' . $id . '"';
+        }
 
-		$label = '<label';
+        if (is_array($attributes) && $attributes) {
+            foreach ($attributes as $key => $val) {
+                $label .= ' ' . $key . '="' . $val . '"';
+            }
+        }
 
-		if ($id !== '')
-		{
-			$label .= ' for="'.$id.'"';
-		}
-
-		if (is_array($attributes) && count($attributes) > 0)
-		{
-			foreach ($attributes as $key => $val)
-			{
-				$label .= ' '.$key.'="'.$val.'"';
-			}
-		}
-
-		return $label.'>'.$label_text.'</label>';
-	}
+        return $label . '>' . $labelText . '</label>';
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_datalist')) {
+    /**
+     * Datalist
+     *
+     * The <datalist> element specifies a list of pre-defined options for an <input> element.
+     * Users will see a drop-down list of pre-defined options as they input data.
+     * The list attribute of the <input> element, must refer to the id attribute of the <datalist> element.
+     */
+    function form_datalist(string $name, string $value, array $options): string
+    {
+        $data = [
+            'type'  => 'text',
+            'name'  => $name,
+            'list'  => $name . '_list',
+            'value' => $value,
+        ];
 
-if ( ! function_exists('form_fieldset'))
-{
-	/**
-	 * Fieldset Tag
-	 *
-	 * Used to produce <fieldset><legend>text</legend>.  To close fieldset
-	 * use form_fieldset_close()
-	 *
-	 * @param	string	The legend text
-	 * @param	array	Additional attributes
-	 * @return	string
-	 */
-	function form_fieldset($legend_text = '', $attributes = array())
-	{
-		$fieldset = '<fieldset'._attributes_to_string($attributes).">\n";
-		if ($legend_text !== '')
-		{
-			return $fieldset.'<legend>'.$legend_text."</legend>\n";
-		}
+        $out = form_input($data) . "\n";
 
-		return $fieldset;
-	}
+        $out .= "<datalist id='" . $name . "_list'>";
+
+        foreach ($options as $option) {
+            $out .= "<option value='{$option}'>\n";
+        }
+
+        return $out . ("</datalist>\n");
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('form_fieldset')) {
+    /**
+     * Fieldset Tag
+     *
+     * Used to produce <fieldset><legend>text</legend>.  To close fieldset
+     * use form_fieldset_close()
+     *
+     * @param string $legendText The legend text
+     * @param array  $attributes Additional attributes
+     */
+    function form_fieldset(string $legendText = '', array $attributes = []): string
+    {
+        $fieldset = '<fieldset' . stringify_attributes($attributes) . ">\n";
 
-if ( ! function_exists('form_fieldset_close'))
-{
-	/**
-	 * Fieldset Close Tag
-	 *
-	 * @param	string
-	 * @return	string
-	 */
-	function form_fieldset_close($extra = '')
-	{
-		return '</fieldset>'.$extra;
-	}
+        if ($legendText !== '') {
+            return $fieldset . '<legend>' . $legendText . "</legend>\n";
+        }
+
+        return $fieldset;
+    }
 }
 
-// ------------------------------------------------------------------------
-
-if ( ! function_exists('form_close'))
-{
-	/**
-	 * Form Close Tag
-	 *
-	 * @param	string
-	 * @return	string
-	 */
-	function form_close($extra = '')
-	{
-		return '</form>'.$extra;
-	}
+if (! function_exists('form_fieldset_close')) {
+    /**
+     * Fieldset Close Tag
+     */
+    function form_fieldset_close(string $extra = ''): string
+    {
+        return '</fieldset>' . $extra;
+    }
 }
 
-// ------------------------------------------------------------------------
-
-if ( ! function_exists('form_prep'))
-{
-	/**
-	 * Form Prep
-	 *
-	 * Formats text so that it can be safely placed in a form field in the event it has HTML tags.
-	 *
-	 * @deprecated	3.0.0	An alias for html_escape()
-	 * @param	string|string[]	$str		Value to escape
-	 * @return	string|string[]	Escaped values
-	 */
-	function form_prep($str)
-	{
-		return html_escape($str, TRUE);
-	}
+if (! function_exists('form_close')) {
+    /**
+     * Form Close Tag
+     */
+    function form_close(string $extra = ''): string
+    {
+        return '</form>' . $extra;
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('set_value')) {
+    /**
+     * Form Value
+     *
+     * Grabs a value from the POST array for the specified field so you can
+     * re-populate an input field or textarea
+     *
+     * @param string          $field      Field name
+     * @param string|string[] $default    Default value
+     * @param bool            $htmlEscape Whether to escape HTML special characters or not
+     *
+     * @return string|string[]
+     */
+    function set_value(string $field, $default = '', bool $htmlEscape = true)
+    {
+        $request = Services::request();
 
-if ( ! function_exists('set_value'))
-{
-	/**
-	 * Form Value
-	 *
-	 * Grabs a value from the POST array for the specified field so you can
-	 * re-populate an input field or textarea. If Form Validation
-	 * is active it retrieves the info from the validation class
-	 *
-	 * @param	string	$field		Field name
-	 * @param	string	$default	Default value
-	 * @param	bool	$html_escape	Whether to escape HTML special characters or not
-	 * @return	string
-	 */
-	function set_value($field, $default = '', $html_escape = TRUE)
-	{
-		$CI =& get_instance();
+        // Try any old input data we may have first
+        $value = $request->getOldInput($field);
 
-		$value = (isset($CI->form_validation) && is_object($CI->form_validation) && $CI->form_validation->has_rule($field))
-			? $CI->form_validation->set_value($field, $default)
-			: $CI->input->post($field, FALSE);
+        if ($value === null) {
+            $value = $request->getPost($field) ?? $default;
+        }
 
-		isset($value) OR $value = $default;
-		return ($html_escape) ? html_escape($value) : $value;
-	}
+        return ($htmlEscape) ? esc($value) : $value;
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('set_select')) {
+    /**
+     * Set Select
+     *
+     * Let's you set the selected value of a <select> menu via data in the POST array.
+     */
+    function set_select(string $field, string $value = '', bool $default = false): string
+    {
+        $request = Services::request();
 
-if ( ! function_exists('set_select'))
-{
-	/**
-	 * Set Select
-	 *
-	 * Let's you set the selected value of a <select> menu via data in the POST array.
-	 * If Form Validation is active it retrieves the info from the validation class
-	 *
-	 * @param	string
-	 * @param	string
-	 * @param	bool
-	 * @return	string
-	 */
-	function set_select($field, $value = '', $default = FALSE)
-	{
-		$CI =& get_instance();
+        // Try any old input data we may have first
+        $input = $request->getOldInput($field);
 
-		if (isset($CI->form_validation) && is_object($CI->form_validation) && $CI->form_validation->has_rule($field))
-		{
-			return $CI->form_validation->set_select($field, $value, $default);
-		}
-		elseif (($input = $CI->input->post($field, FALSE)) === NULL)
-		{
-			return ($default === TRUE) ? ' selected="selected"' : '';
-		}
+        if ($input === null) {
+            $input = $request->getPost($field);
+        }
 
-		$value = (string) $value;
-		if (is_array($input))
-		{
-			// Note: in_array('', array(0)) returns TRUE, do not use it
-			foreach ($input as &$v)
-			{
-				if ($value === $v)
-				{
-					return ' selected="selected"';
-				}
-			}
+        if ($input === null) {
+            return ($default === true) ? ' selected="selected"' : '';
+        }
 
-			return '';
-		}
+        if (is_array($input)) {
+            // Note: in_array('', array(0)) returns TRUE, do not use it
+            foreach ($input as &$v) {
+                if ($value === $v) {
+                    return ' selected="selected"';
+                }
+            }
 
-		return ($input === $value) ? ' selected="selected"' : '';
-	}
+            return '';
+        }
+
+        return ($input === $value) ? ' selected="selected"' : '';
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('set_checkbox')) {
+    /**
+     * Set Checkbox
+     *
+     * Let's you set the selected value of a checkbox via the value in the POST array.
+     */
+    function set_checkbox(string $field, string $value = '', bool $default = false): string
+    {
+        $request = Services::request();
 
-if ( ! function_exists('set_checkbox'))
-{
-	/**
-	 * Set Checkbox
-	 *
-	 * Let's you set the selected value of a checkbox via the value in the POST array.
-	 * If Form Validation is active it retrieves the info from the validation class
-	 *
-	 * @param	string
-	 * @param	string
-	 * @param	bool
-	 * @return	string
-	 */
-	function set_checkbox($field, $value = '', $default = FALSE)
-	{
-		$CI =& get_instance();
+        // Try any old input data we may have first
+        $input = $request->getOldInput($field);
 
-		if (isset($CI->form_validation) && is_object($CI->form_validation) && $CI->form_validation->has_rule($field))
-		{
-			return $CI->form_validation->set_checkbox($field, $value, $default);
-		}
+        if ($input === null) {
+            $input = $request->getPost($field);
+        }
 
-		// Form inputs are always strings ...
-		$value = (string) $value;
-		$input = $CI->input->post($field, FALSE);
+        if (is_array($input)) {
+            // Note: in_array('', array(0)) returns TRUE, do not use it
+            foreach ($input as &$v) {
+                if ($value === $v) {
+                    return ' checked="checked"';
+                }
+            }
 
-		if (is_array($input))
-		{
-			// Note: in_array('', array(0)) returns TRUE, do not use it
-			foreach ($input as &$v)
-			{
-				if ($value === $v)
-				{
-					return ' checked="checked"';
-				}
-			}
+            return '';
+        }
 
-			return '';
-		}
+        $session     = Services::session();
+        $hasOldInput = $session->has('_ci_old_input');
 
-		// Unchecked checkbox and radio inputs are not even submitted by browsers ...
-		if ($CI->input->method() === 'post')
-		{
-			return ($input === $value) ? ' checked="checked"' : '';
-		}
+        // Unchecked checkbox and radio inputs are not even submitted by browsers ...
+        if ((string) $input === '0' || ! empty($request->getPost()) || $hasOldInput) {
+            return ($input === $value) ? ' checked="checked"' : '';
+        }
 
-		return ($default === TRUE) ? ' checked="checked"' : '';
-	}
+        return ($default === true) ? ' checked="checked"' : '';
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('set_radio')) {
+    /**
+     * Set Radio
+     *
+     * Let's you set the selected value of a radio field via info in the POST array.
+     */
+    function set_radio(string $field, string $value = '', bool $default = false): string
+    {
+        $request = Services::request();
 
-if ( ! function_exists('set_radio'))
-{
-	/**
-	 * Set Radio
-	 *
-	 * Let's you set the selected value of a radio field via info in the POST array.
-	 * If Form Validation is active it retrieves the info from the validation class
-	 *
-	 * @param	string	$field
-	 * @param	string	$value
-	 * @param	bool	$default
-	 * @return	string
-	 */
-	function set_radio($field, $value = '', $default = FALSE)
-	{
-		$CI =& get_instance();
+        // Try any old input data we may have first
+        $oldInput = $request->getOldInput($field);
 
-		if (isset($CI->form_validation) && is_object($CI->form_validation) && $CI->form_validation->has_rule($field))
-		{
-			return $CI->form_validation->set_radio($field, $value, $default);
-		}
+        $postInput = $request->getPost($field);
 
-		// Form inputs are always strings ...
-		$value = (string) $value;
-		$input = $CI->input->post($field, FALSE);
+        $input = $oldInput ?? $postInput ?? $default;
 
-		if (is_array($input))
-		{
-			// Note: in_array('', array(0)) returns TRUE, do not use it
-			foreach ($input as &$v)
-			{
-				if ($value === $v)
-				{
-					return ' checked="checked"';
-				}
-			}
+        if (is_array($input)) {
+            // Note: in_array('', array(0)) returns TRUE, do not use it
+            foreach ($input as $v) {
+                if ($value === $v) {
+                    return ' checked="checked"';
+                }
+            }
 
-			return '';
-		}
+            return '';
+        }
 
-		// Unchecked checkbox and radio inputs are not even submitted by browsers ...
-		if ($CI->input->method() === 'post')
-		{
-			return ($input === $value) ? ' checked="checked"' : '';
-		}
+        // Unchecked checkbox and radio inputs are not even submitted by browsers ...
+        if ($oldInput !== null || $postInput !== null) {
+            return ((string) $input === $value) ? ' checked="checked"' : '';
+        }
 
-		return ($default === TRUE) ? ' checked="checked"' : '';
-	}
+        return ($default === true) ? ' checked="checked"' : '';
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('validation_errors')) {
+    /**
+     * Returns the validation errors.
+     *
+     * First, checks the validation errors that are stored in the session.
+     * To store the errors in the session, you need to use `withInput()` with `redirect()`.
+     *
+     * The returned array should be in the following format:
+     *     [
+     *         'field1' => 'error message',
+     *         'field2' => 'error message',
+     *     ]
+     *
+     * @return array<string, string>
+     */
+    function validation_errors()
+    {
+        $errors = session('_ci_validation_errors');
 
-if ( ! function_exists('form_error'))
-{
-	/**
-	 * Form Error
-	 *
-	 * Returns the error for a specific form field. This is a helper for the
-	 * form validation class.
-	 *
-	 * @param	string
-	 * @param	string
-	 * @param	string
-	 * @return	string
-	 */
-	function form_error($field = '', $prefix = '', $suffix = '')
-	{
-		if (FALSE === ($OBJ =& _get_validation_object()))
-		{
-			return '';
-		}
+        // Check the session to see if any were
+        // passed along from a redirect withErrors() request.
+        if ($errors !== null && (ENVIRONMENT === 'testing' || ! is_cli())) {
+            return $errors;
+        }
 
-		return $OBJ->error($field, $prefix, $suffix);
-	}
+        $validation = Services::validation();
+
+        return $validation->getErrors();
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('validation_list_errors')) {
+    /**
+     * Returns the rendered HTML of the validation errors.
+     *
+     * See Validation::listErrors()
+     */
+    function validation_list_errors(string $template = 'list'): string
+    {
+        $config = config(Validation::class);
+        $view   = Services::renderer();
 
-if ( ! function_exists('validation_errors'))
-{
-	/**
-	 * Validation Error String
-	 *
-	 * Returns all the errors associated with a form submission. This is a helper
-	 * function for the form validation class.
-	 *
-	 * @param	string
-	 * @param	string
-	 * @return	string
-	 */
-	function validation_errors($prefix = '', $suffix = '')
-	{
-		if (FALSE === ($OBJ =& _get_validation_object()))
-		{
-			return '';
-		}
+        if (! array_key_exists($template, $config->templates)) {
+            throw ValidationException::forInvalidTemplate($template);
+        }
 
-		return $OBJ->error_string($prefix, $suffix);
-	}
+        return $view->setVar('errors', validation_errors())
+            ->render($config->templates[$template]);
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('validation_show_error')) {
+    /**
+     * Returns a single error for the specified field in formatted HTML.
+     *
+     * See Validation::showError()
+     */
+    function validation_show_error(string $field, string $template = 'single'): string
+    {
+        $config = config(Validation::class);
+        $view   = Services::renderer();
 
-if ( ! function_exists('_parse_form_attributes'))
-{
-	/**
-	 * Parse the form attributes
-	 *
-	 * Helper function used by some of the form helpers
-	 *
-	 * @param	array	$attributes	List of attributes
-	 * @param	array	$default	Default values
-	 * @return	string
-	 */
-	function _parse_form_attributes($attributes, $default)
-	{
-		if (is_array($attributes))
-		{
-			foreach ($default as $key => $val)
-			{
-				if (isset($attributes[$key]))
-				{
-					$default[$key] = $attributes[$key];
-					unset($attributes[$key]);
-				}
-			}
+        $errors = array_filter(validation_errors(), static fn ($key) => preg_match(
+            '/^' . str_replace(['\.\*', '\*\.'], ['\..+', '.+\.'], preg_quote($field, '/')) . '$/',
+            $key
+        ), ARRAY_FILTER_USE_KEY);
 
-			if (count($attributes) > 0)
-			{
-				$default = array_merge($default, $attributes);
-			}
-		}
+        if ($errors === []) {
+            return '';
+        }
 
-		$att = '';
+        if (! array_key_exists($template, $config->templates)) {
+            throw ValidationException::forInvalidTemplate($template);
+        }
 
-		foreach ($default as $key => $val)
-		{
-			if ($key === 'value')
-			{
-				$val = html_escape($val);
-			}
-			elseif ($key === 'name' && ! strlen($default['name']))
-			{
-				continue;
-			}
-
-			$att .= $key.'="'.$val.'" ';
-		}
-
-		return $att;
-	}
+        return $view->setVar('error', implode("\n", $errors))
+            ->render($config->templates[$template]);
+    }
 }
 
-// ------------------------------------------------------------------------
+if (! function_exists('parse_form_attributes')) {
+    /**
+     * Parse the form attributes
+     *
+     * Helper function used by some of the form helpers
+     *
+     * @internal
+     *
+     * @param array|string $attributes List of attributes
+     * @param array        $default    Default values
+     */
+    function parse_form_attributes($attributes, array $default): string
+    {
+        if (is_array($attributes)) {
+            foreach (array_keys($default) as $key) {
+                if (isset($attributes[$key])) {
+                    $default[$key] = $attributes[$key];
+                    unset($attributes[$key]);
+                }
+            }
+            if ($attributes !== []) {
+                $default = array_merge($default, $attributes);
+            }
+        }
 
-if ( ! function_exists('_attributes_to_string'))
-{
-	/**
-	 * Attributes To String
-	 *
-	 * Helper function used by some of the form helpers
-	 *
-	 * @param	mixed
-	 * @return	string
-	 */
-	function _attributes_to_string($attributes)
-	{
-		if (empty($attributes))
-		{
-			return '';
-		}
+        $att = '';
 
-		if (is_object($attributes))
-		{
-			$attributes = (array) $attributes;
-		}
+        foreach ($default as $key => $val) {
+            if (! is_bool($val)) {
+                if ($key === 'value') {
+                    $val = esc($val);
+                } elseif ($key === 'name' && ! strlen($default['name'])) {
+                    continue;
+                }
+                $att .= $key . '="' . $val . '"' . ($key === array_key_last($default) ? '' : ' ');
+            } else {
+                $att .= $key . ' ';
+            }
+        }
 
-		if (is_array($attributes))
-		{
-			$atts = '';
-
-			foreach ($attributes as $key => $val)
-			{
-				$atts .= ' '.$key.'="'.$val.'"';
-			}
-
-			return $atts;
-		}
-
-		if (is_string($attributes))
-		{
-			return ' '.$attributes;
-		}
-
-		return FALSE;
-	}
-}
-
-// ------------------------------------------------------------------------
-
-if ( ! function_exists('_get_validation_object'))
-{
-	/**
-	 * Validation Object
-	 *
-	 * Determines what the form validation class was instantiated as, fetches
-	 * the object and returns it.
-	 *
-	 * @return	mixed
-	 */
-	function &_get_validation_object()
-	{
-		$CI =& get_instance();
-
-		// We set this as a variable since we're returning by reference.
-		$return = FALSE;
-
-		if (FALSE !== ($object = $CI->load->is_loaded('Form_validation')))
-		{
-			if ( ! isset($CI->$object) OR ! is_object($CI->$object))
-			{
-				return $return;
-			}
-
-			return $CI->$object;
-		}
-
-		return $return;
-	}
+        return $att;
+    }
 }
